requirements.txt
uvicorn
fastapi
pyyaml
requests